Refactor constants and networks
