+++ src/utils/constants.js
@@ -1,0 +1,9 @@
+const zeroAddress = `0x${'0'.repeat(40)}`;
+
+// keccak256(toUtf8Bytes('Contract Proxy Kit'))
+const predeterminedSaltNonce = '0xcfe33a586323e7325be6aa6ecd8b4600d232a9037e83c8ece69413b777dabe65';
+
+module.exports = {
+  zeroAddress,
+  predeterminedSaltNonce
+}
